Ltoloza/refactor/domain hardening (#33)
* refactor(stock): refactor code to be able to use of a BigDecimal and a barcode object for improve the scalability.

* test(stock): refactor unitest for make them able to verify the refactored code

* fix(domain): add rounding mode to BigDecimal division to satisfy Qodana

---------

+++ src/main/java/domain/configuration/GlobalConfig.java
@@ -4,35 +4,32 @@ import domain.common.BaseEntity;
 import domain.common.Guard;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
+import java.math.BigDecimal;
 
-/**
- * Represents the system-wide configuration.
- * Intended to be a singleton (only one record in the database).
- */
 @Entity
 @Table(name = "global_configuration")
 public class GlobalConfig extends BaseEntity {
 
-    private double defaultProfitMargin;
+    private BigDecimal defaultProfitMargin;
 
     protected GlobalConfig() {
     }
 
-    public GlobalConfig(double defaultProfitMargin) {
+    public GlobalConfig(BigDecimal defaultProfitMargin) {
         setProfitMargin(defaultProfitMargin);
     }
 
-    public void updateProfitMargin(double newMargin) {
+    public void updateProfitMargin(BigDecimal newMargin) {
         setProfitMargin(newMargin);
     }
 
-    private void setProfitMargin(double margin) {
+    private void setProfitMargin(BigDecimal margin) {
         Guard.againstNegative(margin, "Profit Margin");
         Guard.againstUnrealisticProfitMargin(margin, "Profit Margin");
         this.defaultProfitMargin = margin;
     }
 
-    public double getDefaultProfitMargin() {
+    public BigDecimal getDefaultProfitMargin() {
         return defaultProfitMargin;
     }
 }
